// src/app/models/Mentor/mentor.interface.ts
export interface  mentorData {
    mentorId: string;
    name: string;
    gender: 'male' | 'female' | 'other';
    dateOfBirth: string;
    email: string;
    contactNo: string;
    emergencyContact: string;
    addrees: string;
    profileImg: string;
    designation: string;
    departmentName: string;
    specialized_area: string;
    education_qualification: string[];
    workExperience:string;
    experienceYears: string;
    experienceTrainedStudents: string;
    reviews: number;
    bio: string;
    lifeJounery: string;
}                                                                           